Add files via upload
In our portfolio project, we've leveraged HTML and CSS to create a visually stunning and functionally rich website that showcases your skills, projects, and achievements as a web developer. Here's a brief overview of what we've accomplished:

HTML Structure: We've crafted a well-organized HTML structure that lays the foundation for our portfolio website. This structure includes various sections such as the header, profile summary, education, technical skills, projects, internships, skills, extracurricular activities, and a declaration.
CSS Styling: Using CSS, we've applied styling to enhance the visual appeal and user experience of the website. This includes defining colors, fonts, spacing, borders, backgrounds, and other visual elements to create a cohesive and attractive design.